Add skeleton for Bosh release
Add most files required to package, configure and run the CDN broker.

In particular, the spec file provides a way of adding the necessary
environment variables for configuration of the broker. We must be
cautious with the variable names as the broker uses a
library which looks for environment variables starting with a prefix
('CDN') and end with a name derived from the `Settings` struct in the
source code.