Supply the catalog.json file
The broker hard codes the location of the catalog.json file[1],
so we have to copy it to the `pwd` of the running process until a
better solution can be found. Ideally, we should pass the catalog details
via the manifest and extend the broker so it can be configured to look
for the catalog in a different location.

[1]
https://github.com/18F/cf-cdn-service-broker/blob/672c141b84d8ee1f05a7ab125534cd797d32db24/broker/broker.go#L50