Use our fork of the CDN broker
We have to make updates to the broker to add a non-basic-auth health
check endpoint for the ELB.

This points the submodule to our fork, but keeps it under the same
path as the upstream repository (github.com/18F/cf-cdn-service-broker),
so import paths are not broken. For developing locally you can checkout
this repository, update the submodules, and add the upstream repository
as a remote.

We may revert to the upstream repo once our changes are merged.