Add cdn-cron job
Add the CDN cron process (a Golang process that uses a cron-like library)
to manage certificate renewals.

We have duplicated much of the configuration (the job properties shell script and spec
file properties) because the cron process shares the same `config`
package[1][2] and a lot of the fields are required, regardless of the fact
the cron process doesn't need them.

We use the special `spec.bootstrap` property[3] in the Monit template to ensure
Monit is only configured on one VM. Otherwise we would have race conditions
in the renewal of certificates and could unnecessarily use up our Let's Encrypt
limits.

[1] https://github.com/18F/cf-cdn-service-broker/blob/fb84788e8e74652c2748c06f153d5c404abe7ef2/cmd/cdn-cron/main.go#L15
[2] https://github.com/18F/cf-cdn-service-broker/blob/fb84788e8e74652c2748c06f153d5c404abe7ef2/config/config.go#L11-L30
[3] https://bosh.io/docs/jobs.html#properties